+++ testsuite/test_nocasedict.py
@@ -6,6 +6,8 @@
 from __future__ import absolute_import
 
 import unittest
+import warnings
+import six
 
 from pywbem.cim_obj import NocaseDict
 
@@ -246,59 +248,165 @@ class TestPopItem(BaseTest):
     def test_all(self):
         pass
 
+# TODO: swapcase2() is also defined in test_cim_obj.py. Consolidate.
+def swapcase2(text):
+    """Returns text, where every other character has been changed to swap
+    its lexical case. For strings that contain at least one letter, the
+    returned string is guaranteed to be different from the input string."""
+    text_cs = ''
+    i = 0
+    for c in text:
+        if i % 2 != 0:
+            c = c.swapcase()
+        text_cs += c
+        i += 1
+    return text_cs
+
 class TestEqual(BaseTest):
 
-    def test_all(self):
-        dic2 = NocaseDict({'dog': 'Cat', 'Budgie': 'Fish'})
-        self.assertTrue(self.dic == dic2)
-        dic2['Budgie'] = 'fish'
-        self.assertTrue(self.dic != dic2)
+    def assertDictEqual(self, dic1, dic2, msg):
 
-class TestComparison(BaseTest):
+        self.assertTrue(dic1 == dic2, msg)
+        self.assertFalse(dic1 != dic2, msg)
 
-    def assertSame(self, dic1, dic2):
+        self.assertTrue(dic2 == dic1, msg)
+        self.assertFalse(dic2 != dic1, msg)
 
-        self.assertTrue(dic1 == dic2)
-        self.assertFalse(dic1 != dic2)
-        self.assertTrue(dic1 >= dic2)
-        self.assertFalse(dic1 > dic2)
-        self.assertTrue(dic1 <= dic2)
-        self.assertFalse(dic1 < dic2)
+    def assertDictNotEqual(self, dic1, dic2, msg):
 
-        self.assertTrue(dic2 == dic1)
-        self.assertFalse(dic2 != dic1)
-        self.assertTrue(dic2 >= dic1)
-        self.assertFalse(dic2 > dic1)
-        self.assertTrue(dic2 <= dic1)
-        self.assertFalse(dic2 < dic1)
+        self.assertTrue(dic1 != dic2, msg)
+        self.assertFalse(dic1 == dic2, msg)
 
-    def assertLess(self, dic1, dic2):
+        self.assertTrue(dic2 != dic1, msg)
+        self.assertFalse(dic2 == dic1, msg)
 
-        self.assertTrue(dic1 != dic2)
-        self.assertFalse(dic1 == dic2)
-        self.assertTrue(dic1 < dic2)
-        self.assertFalse(dic1 > dic2)
-        self.assertTrue(dic1 <= dic2)
-        self.assertFalse(dic1 >= dic2)
+    def run_test_dicts(self, base_dict, test_dicts):
 
-        self.assertTrue(dic2 != dic1)
-        self.assertFalse(dic2 == dic1)
-        self.assertTrue(dic2 > dic1)
-        self.assertFalse(dic2 < dic1)
-        self.assertTrue(dic2 >= dic1)
-        self.assertFalse(dic2 <= dic1)
+        for test_dict, relation, comment in test_dicts:
+            if relation == 'eq':
+                self.assertDictEqual(test_dict, base_dict,
+                                "Expected test_dict == base_dict:\n" \
+                                "  test case: %s\n" \
+                                "  test_dict: %r\n" \
+                                "  base_dict: %r" % \
+                                (comment, test_dict, base_dict))
+            elif relation == 'ne':
+                self.assertDictNotEqual(test_dict, base_dict,
+                                "Expected test_dict != base_dict:\n" \
+                                "  test case: %s\n" \
+                                "  test_dict: %r\n" \
+                                "  base_dict: %r" % \
+                                (comment, test_dict, base_dict))
+            else:
+                raise AssertionError("Internal Error: Invalid relation %s" \
+                                     "specified in testcase: %s" % \
+                                     (relation, comment))
 
     def test_all(self):
-        dic_same = NocaseDict({'doG': 'Cat', 'BuDgie': 'Fish'})
-        dic_less1 = NocaseDict({'doG': 'Cat'})
-        dic_less2 = NocaseDict({'DOg': 'Cat', 'Alf': 'Horse'})
-        dic_less3 = NocaseDict({'doG': 'Car', 'budGie': 'Fish'})
 
-        self.assertSame(dic_same, self.dic)
-        # TODO: Enable these tests to work on the dict ordering issue
-        #self.assertLess(dic_less1, self.dic)
-        #self.assertLess(dic_less2, self.dic)
-        #self.assertLess(dic_less3, self.dic)
+        # The base dictionary that is used for all comparisons
+        base_dict = dict({'Budgie': 'Fish', 'Dog': 'Cat'})
+
+        # Test dictionaries to test against the base dict, as a list of
+        # tuple(dict, relation, comment), with relation being the expected
+        # comparison relation, and one of ('eq', 'ne').
+        test_dicts = [
+
+            (dict({'Budgie': 'Fish', 'Dog': 'Cat'}),
+             'eq',
+             'Same'),
+
+            (dict({'Budgie': 'Fish'}),
+             'ne',
+             'Higher key missing, shorter size'),
+
+            (dict({'Dog': 'Cat'}),
+             'ne',
+             'Lower key missing, shorter size'),
+
+            (dict({'Budgie': 'Fish', 'Curly': 'Snake', 'Cozy': 'Dog'}),
+             'ne',
+             'First non-matching key is less. But longer size!'),
+
+            (dict({'Alf': 'F', 'Anton': 'S', 'Aussie': 'D'}),
+             'ne',
+             'Only non-matching keys that are less. But longer size!'),
+
+            (dict({'Budgio': 'Fish'}),
+             'ne',
+             'First non-matching key is greater. But shorter size!'),
+
+            (dict({'Zoe': 'F'}),
+             'ne',
+             'Only non-matching keys that are greater. But shorter size!'),
+
+            (dict({'Budgie': 'Fish', 'Curly': 'Snake'}),
+             'ne',
+             'Same size. First non-matching key is less'),
+
+            (dict({'Alf': 'F', 'Anton': 'S'}),
+             'ne',
+             'Same size. Only non-matching keys that are less'),
+
+            (dict({'Zoe': 'F', 'Zulu': 'S'}),
+             'ne',
+             'Same size. Only non-matching keys that are greater'),
+
+            (dict({'Budgie': 'Fish', 'Dog': 'Car'}),
+             'ne',
+             'Same size, only matching keys. First non-matching value is less'),
+
+            (dict({'Budgie': 'Fish', 'Dog': 'Caz'}),
+             'ne',
+             'Same size, only matching keys. First non-matching value is grt.'),
+        ]
+
+        # First, run these tests against a standard dictionary to verify
+        # that the test case definitions conform to that
+        self.run_test_dicts(base_dict, test_dicts)
+
+        # Then, transform these tests to NocaseDict and run them again
+        TEST_CASE_INSENSITIVITY = True
+        base_ncdict = NocaseDict(base_dict)
+        test_ncdicts = []
+        for test_dict, relation, comment in test_dicts:
+            test_ncdict = NocaseDict()
+            for key in test_dict:
+                if TEST_CASE_INSENSITIVITY:
+                    nc_key = swapcase2(key)
+                else:
+                    nc_key = key
+                test_ncdict[nc_key] = test_dict[key]
+            test_ncdicts.append((test_ncdict, relation, comment))
+        self.run_test_dicts(base_ncdict, test_ncdicts)
+
+class TestOrdering(BaseTest):
+    """Verify that ordering comparisons between NocaseDict instances
+    issue a deprecation warning, and for Python 3, in addition the usual
+    "TypeError: unorderable types" for standard dicts."""
+
+    def assertWarning(self, comp_str):
+        with warnings.catch_warnings(record=True) as wlist:
+            warnings.simplefilter("always")
+            if six.PY2:
+                eval(comp_str)
+            else:
+                try:
+                    eval(comp_str)
+                except TypeError as exc:
+                    assert "unorderable types" in str(exc)
+                else:
+                    self.fail("Ordering a dictionary in Python 3 did not "
+                              "raise TypeError")
+            assert len(wlist) >= 1
+            assert issubclass(wlist[-1].category, DeprecationWarning)
+            assert "deprecated" in str(wlist[-1].message)
+
+    def test_all(self):
+        self.assertWarning("self.dic < self.dic")
+        self.assertWarning("self.dic <= self.dic")
+        self.assertWarning("self.dic > self.dic")
+        self.assertWarning("self.dic >= self.dic")
 
 class TestContains(BaseTest):